refactor(trembita-cli): simplify argument handling and improve showcase resolution
- Refactored argument retrieval in the `http.rs` and `mod.rs` files to remove unnecessary references, enhancing code clarity.
- Updated the `listen_addr` method in `showcases.rs` to streamline port calculation.
- Cleaned up unused imports in `markers.rs`, improving overall code quality and maintainability.

+++ crates/trembita-cli/src/dev/showcases.rs
@@ -140,7 +140,7 @@ impl Showcase {
     /// `127.0.0.1:<port>` for node index `1..=nodes`.
     #[must_use]
     pub fn listen_addr(&self, node: u32) -> String {
-        let port = u32::from(self.base_port) + u32::from(node.saturating_sub(1));
+        let port = u32::from(self.base_port) + node.saturating_sub(1);
         format!("127.0.0.1:{port}")
     }
